+++ src/storage.rs
@@ -40,20 +40,20 @@ pub use store_prefix::{StorePrefix, StorePrefixError, StorePrefixes};
 #[cfg(feature = "async")]
 pub use self::storage_async::{
     async_create_array, async_create_group, async_discover_children, async_discover_nodes,
-    async_erase_chunk, async_erase_node, async_get_child_nodes, async_node_exists,
-    async_node_exists_listable, async_retrieve_chunk, async_retrieve_partial_values,
-    async_store_chunk, async_store_set_partial_values, AsyncListableStorageTraits,
-    AsyncReadableListableStorageTraits, AsyncReadableStorageTraits,
+    async_erase_chunk, async_erase_metadata, async_erase_node, async_get_child_nodes,
+    async_node_exists, async_node_exists_listable, async_retrieve_chunk,
+    async_retrieve_partial_values, async_store_chunk, async_store_set_partial_values,
+    AsyncListableStorageTraits, AsyncReadableListableStorageTraits, AsyncReadableStorageTraits,
     AsyncReadableWritableListableStorageTraits, AsyncReadableWritableStorageTraits,
     AsyncWritableStorageTraits,
 };
 
 pub use self::storage_sync::{
-    create_array, create_group, discover_children, discover_nodes, erase_chunk, erase_node,
-    get_child_nodes, node_exists, node_exists_listable, retrieve_chunk, retrieve_partial_values,
-    store_chunk, store_set_partial_values, ListableStorageTraits, ReadableListableStorageTraits,
-    ReadableStorageTraits, ReadableWritableListableStorageTraits, ReadableWritableStorageTraits,
-    WritableStorageTraits,
+    create_array, create_group, discover_children, discover_nodes, erase_chunk, erase_metadata,
+    erase_node, get_child_nodes, node_exists, node_exists_listable, retrieve_chunk,
+    retrieve_partial_values, store_chunk, store_set_partial_values, ListableStorageTraits,
+    ReadableListableStorageTraits, ReadableStorageTraits, ReadableWritableListableStorageTraits,
+    ReadableWritableStorageTraits, WritableStorageTraits,
 };
 pub use self::storage_transformer::StorageTransformerChain;